api/pyproject.toml
[project]
dependencies = [
    "fastapi>=0.117.1",
    "uvicorn>=0.36.0",
    "requests>=2.32.3",
    'python-jose[cryptography]>=3.3.0',
    "python-dotenv>=1.0.0",
]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
